+++ NEWS.md
@@ -1,6 +1,7 @@
 # tinysnapshot 0.0.2.9000
 
 * New `os` argument to skip tests on unspecified operating systems.
+* New `device_args` argument to pass additional arguments to the device function.
 
 # tinysnapshot 0.0.2